+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/adapter/PolarisConfigCustomExtensionLayer.java
@@ -31,8 +31,9 @@ import org.springframework.core.env.Environment;
 public interface PolarisConfigCustomExtensionLayer {
 	void initRegisterConfig(PolarisConfigPropertyAutoRefresher polarisConfigPropertyAutoRefresher);
 
-	void initConfigFiles(Environment environment, CompositePropertySource compositePropertySource, PolarisPropertySourceManager polarisPropertySourceManager, ConfigFileService configFileService);
+	void initConfigFiles(Environment environment, CompositePropertySource compositePropertySource, ConfigFileService configFileService);
 
 	void executeAfterLocateConfigReturning(CompositePropertySource compositePropertySource);
+
 	boolean executeRegisterPublishChangeListener(PolarisPropertySource polarisPropertySource);
 }